backend_app/requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
coverage==4.5.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-money==0.15.1
django-rest-auth==0.9.5
django-stubs==1.5.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.1.0
Faker==0.9.1
httpie==1.0.3
idna==2.8
importlib-metadata==0.19
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==3.1.1
mixer==6.1.3
more-itertools==7.2.0
mypy==0.770
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==19.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
py-moneyed==0.8.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
